refactor(reset): add reset functions to deployer interface
Start reset refactor by integrating the main functions in Deployer
interface and separating again the utils functions in another file.
