+++ internal/command/admin/reset.go
@@ -20,8 +20,6 @@ package admin
 import (
 	"context"
 
-	"github.com/docker/docker/client"
-	"github.com/funlessdev/fl-cli/pkg/deploy"
 	"github.com/funlessdev/fl-cli/pkg/log"
 )
 
@@ -30,37 +28,37 @@ type reset struct{}
 func (r *reset) Run(ctx context.Context, logger log.FLogger) error {
 	logger.Info("Removing local funless deployment...\n")
 
-	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithVersion("1.41"))
-	if err != nil {
-		return err
-	}
+	// cli, err := client.NewClientWithOpts(client.FromEnv, client.WithVersion("1.41"))
+	// if err != nil {
+	// 	return err
+	// }
 
-	_ = logger.StartSpinner("Removing Core container... ☠️")
-	if err := logger.StopSpinner(deploy.RemoveFLContainer(ctx, cli, "fl-core")); err != nil {
-		return err
-	}
+	// _ = logger.StartSpinner("Removing Core container... ☠️")
+	// if err := logger.StopSpinner(deploy.RemoveFLContainer(ctx, cli, "fl-core")); err != nil {
+	// 	return err
+	// }
 
-	_ = logger.StartSpinner("Removing Worker container... 🔪")
-	if err := logger.StopSpinner(deploy.RemoveFLContainer(ctx, cli, "fl-worker")); err != nil {
-		return err
-	}
+	// _ = logger.StartSpinner("Removing Worker container... 🔪")
+	// if err := logger.StopSpinner(deploy.RemoveFLContainer(ctx, cli, "fl-worker")); err != nil {
+	// 	return err
+	// }
 
-	_ = logger.StartSpinner("Removing the function containers... 🔫")
-	if err := logger.StopSpinner(deploy.RemoveFunctionContainers(ctx, cli)); err != nil {
-		return err
-	}
+	// _ = logger.StartSpinner("Removing the function containers... 🔫")
+	// if err := logger.StopSpinner(deploy.RemoveFunctionContainers(ctx, cli)); err != nil {
+	// 	return err
+	// }
 
-	_ = logger.StartSpinner("Removing fl_net network... ✂️")
-	if err := logger.StopSpinner(deploy.RemoveFLNetwork(ctx, cli, "fl_net")); err != nil {
-		return err
-	}
+	// _ = logger.StartSpinner("Removing fl_net network... ✂️")
+	// if err := logger.StopSpinner(deploy.RemoveFLNetwork(ctx, cli, "fl_net")); err != nil {
+	// 	return err
+	// }
 
-	_ = logger.StartSpinner("Removing fl_runtime_net network... ✂️")
-	if err := logger.StopSpinner(deploy.RemoveFLNetwork(ctx, cli, "fl_runtime_net")); err != nil {
-		return err
-	}
+	// _ = logger.StartSpinner("Removing fl_runtime_net network... ✂️")
+	// if err := logger.StopSpinner(deploy.RemoveFLNetwork(ctx, cli, "fl_runtime_net")); err != nil {
+	// 	return err
+	// }
 
 	logger.Info("\nAll clear! 👍")
 
-	return err
+	return nil
 }
