+++ pkg/deploy/types.go
@@ -21,9 +21,16 @@ import "context"
 
 type DockerDeployer interface {
 	SetupClient(ctx context.Context) error
+
 	SetupFLNetworks(ctx context.Context) error
 	PullCoreImage(ctx context.Context) error
 	PullWorkerImage(ctx context.Context) error
 	StartCore(ctx context.Context) error
 	StartWorker(ctx context.Context) error
+
+	RemoveFLMainNetwork(ctx context.Context) error
+	RemoveFLRuntimeNetwork(ctx context.Context) error
+	RemoveCoreContainer(context.Context) error
+	RemoveWorkerContainer(context.Context) error
+	RemoveFunctionContainers(ctx context.Context) error
 }
